InlineSecureValue: Move to its own package rather than service
